Make rebuilding customizable and avoid catching all exceptions
There are several reasons why `frog.rkt` might not be loadable. We
only want to show a nice warning that users should initialize the
project when the exception is filesystem-related. Other errors like
syntax errors / unbound ids should not be caught.

Improve --watch flag by allowing users to customize whether or not
they want to trigger a rebuild or not. The default behavior is
backward-compatible with the current behavior.

Make --watch withstands exceptions. Beep without new line.